requirements.txt
# API and Web Framework
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
httpx>=0.24.1

# LLM Integration
groq==0.3.0

# Environment Variables
python-dotenv==1.0.0

# HTTP Client
requests==2.31.0

# Date and Time
python-dateutil==2.8.2

# Type Hints
typing-extensions==4.9.0

# Optional: Testing
pytest==8.0.0
pytest-asyncio==0.23.5
